+++ massiv/src/Data/Massiv/Array/Delayed.hs
@@ -11,6 +11,7 @@ module Data.Massiv.Array.Delayed
   -- ** Delayed Pull Array
     D(..)
   , delay
+  , liftArray2M
   -- ** Delayed Push Array
   , DL(..)
   , toLoadArray